+++ src/interpreter/lib/std.ts
@@ -276,6 +276,29 @@ export const std: Record<string, Value> = {
 
 	//#region Str
 	'Str:lf': STR('\n'),
+
+	'Str:lt': FN_NATIVE(([a, b]) => {
+		assertString(a);
+		assertString(b);
+		if (a.value < b.value) {
+			return NUM(-1);
+		}else if (a.value === b.value){
+			return NUM(0);
+		}else {
+			return NUM(1);
+		}
+	}),
+	'Str:gt': FN_NATIVE(([a, b]) => {
+		assertString(a);
+		assertString(b);
+		if (a.value > b.value) {
+			return NUM(-1);
+		}else if (a.value === b.value){
+			return NUM(0);
+		}else {
+			return NUM(1);
+		}
+	}),
 	//#endregion
 
 	//#region Arr